# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API BackEnd",
    "version": "1.0.0",
    "description": "Documentação para a criação de usuários"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/atividades/types": {
      "summary": "Operações relacionadas às atividades",
      "description": "Criação e consulta de atividades no sistema",
      "tags": ["Atividades"],
      "get": {
        "summary": "Busca todos os tipos de atividade",
        "description": "Retorna todos os tipos de atividades cadastrados no sistema.",
        "operationId": "getAllTypes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Atividades"],
        "responses": {
          "200": {
            "description": "Tipos de atividades encontrados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID do tipo de atividade"
                      },
                      "name": {
                        "type": "string",
                        "description": "Nome do tipo de atividade"
                      },
                      "description": {
                        "type": "string",
                        "description": "Descrição do tipo de atividade"
                      },
                      "image": {
                        "type": "string",
                        "description": "URL da imagem do tipo de atividade"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhum tipo de atividade encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "post": {
        "summary": "Cria uma nova atividade",
        "description": "Endpoint para criar uma nova atividade no sistema",
        "operationId": "createActivity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Atividades"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Título da atividade"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição da atividade"
                  },
                  "typeId": {
                    "type": "string",
                    "description": "ID do tipo de atividade"
                  },
                  "scheduledDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data e hora agendada para a atividade"
                  },
                  "image": {
                    "type": "string",
                    "description": "URL da imagem associada à atividade"
                  },
                  "private": {
                    "type": "boolean",
                    "description": "Indica se a atividade é privada"
                  },
                  "creatorId": {
                    "type": "string",
                    "description": "ID do criador da atividade"
                  },
                  "latitude": {
                    "type": "number",
                    "format": "float",
                    "description": "Latitude do endereço da atividade"
                  },
                  "longitude": {
                    "type": "number",
                    "format": "float",
                    "description": "Longitude do endereço da atividade"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "typeId",
                  "scheduledDate",
                  "image",
                  "private",
                  "creatorId",
                  "latitude",
                  "longitude"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atividade criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID da atividade"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título da atividade"
                    },
                    "description": {
                      "type": "string",
                      "description": "Descrição da atividade"
                    },
                    "type": {
                      "type": "string",
                      "description": "Tipo da atividade"
                    },
                    "image": {
                      "type": "string",
                      "description": "URL da imagem associada à atividade"
                    },
                    "address": {
                      "type": "object",
                      "properties": {
                        "latitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Latitude do endereço"
                        },
                        "longitude": {
                          "type": "number",
                          "format": "float",
                          "description": "Longitude do endereço"
                        }
                      }
                    },
                    "scheduledDate": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Data e hora agendada para a atividade"
                    },
                    "private": {
                      "type": "boolean",
                      "description": "Indica se a atividade é privada"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Data de criação da atividade"
                    },
                    "completedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Data de conclusão da atividade"
                    },
                    "creator": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "ID do criador"
                        },
                        "name": {
                          "type": "string",
                          "description": "Nome do criador"
                        },
                        "avatar": {
                          "type": "string",
                          "description": "URL do avatar do criador"
                        }
                      }
                    },
                    "userSubscriptionStatus": {
                      "type": "string",
                      "description": "Status da inscrição do usuário"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/atividades/{id}": {
      "put": {
        "summary": "Atualiza uma atividade",
        "description": "Atualiza os dados de uma atividade existente.",
        "tags": ["Atividades"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da atividade a ser atualizada"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "typeId": { "type": "string" },
                  "image": { "type": "string" },
                  "latitude": { "type": "number", "format": "float" },
                  "longitude": { "type": "number", "format": "float" },
                  "scheduledDate": { "type": "string", "format": "date-time" },
                  "privateActivity": { "type": "boolean" },
                  "creatorId": { "type": "string" },
                  "completedAt": { "type": "string", "format": "date-time" }
                },
                "required": ["title", "description", "typeId", "scheduledDate"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atividade atualizada com sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "404": {
            "description": "Atividade não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Exclui uma atividade",
        "description": "Exclui uma atividade existente no sistema.",
        "tags": ["Atividades"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da atividade a ser excluída"
          }
        ],
        "responses": {
          "200": {
            "description": "Atividade excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de confirmação de exclusão"
                    },
                    "deletedActivity": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "ID da atividade"
                        },
                        "title": {
                          "type": "string",
                          "description": "Título da atividade"
                        },
                        "description": {
                          "type": "string",
                          "description": "Descrição da atividade"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Atividade não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/atividades/{id}/subscribe": {
      "post": {
        "summary": "Inscrever-se em uma atividade",
        "description": "Endpoint para inscrever-se em uma atividade. O usuário deve fornecer seu ID para se inscrever na atividade especificada.",
        "operationId": "subscribeToActivity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da atividade em que o usuário deseja se inscrever"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "ID do usuário que deseja se inscrever na atividade"
                  }
                },
                "required": ["userId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inscrição realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de sucesso da inscrição"
                    },
                    "subscription": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "ID da inscrição"
                        },
                        "userId": {
                          "type": "string",
                          "description": "ID do usuário"
                        },
                        "activityId": {
                          "type": "string",
                          "description": "ID da atividade"
                        },
                        "approved": {
                          "type": "boolean",
                          "description": "Status da aprovação da inscrição"
                        },
                        "confirmedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Data e hora de confirmação da inscrição"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Atividade não encontrada"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/atividades/participant/all": {
      "summary": "Busca todas as atividades de um usuário participante",
      "description": "Retorna todas as atividades relacionadas a um usuário autenticado no sistema.",
      "tags": ["Atividades"],
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "get": {
        "summary": "Busca atividades de um usuário",
        "description": "Este endpoint retorna todas as atividades associadas ao usuário autenticado.",
        "operationId": "getUserActivities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Atividades"],
        "responses": {
          "200": {
            "description": "Atividades do usuário encontradas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID da atividade"
                      },
                      "title": {
                        "type": "string",
                        "description": "Título da atividade"
                      },
                      "description": {
                        "type": "string",
                        "description": "Descrição da atividade"
                      },
                      "scheduledDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Data e hora agendada para a atividade"
                      },
                      "image": {
                        "type": "string",
                        "description": "URL da imagem associada à atividade"
                      },
                      "private": {
                        "type": "boolean",
                        "description": "Indica se a atividade é privada"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhuma atividade encontrada para o usuário"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  }
}